requirements.txt
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.1.0
PyMuPDF==1.23.8
reportlab==4.0.9
